+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/sunng87/handlebars-rust/compare/3.0.1...Unreleased) - ReleaseDate
 
 * [Added] multiple parameter support for log helper
+* [Fixed] helper lookup priority
 
 ## [3.0.1](https://github.com/sunng87/handlebars-rust/compare/3.0.0...3.0.1) - 2020-01-25